Extract createWorkpieceSlice from projectStore (P5 store slice)
