+++ src/store/projectStore.ts
@@ -39,9 +39,7 @@ import {
   inferFeatureKind,
   newProject,
   profileVertices,
-  rectProfile,
   circleProfile,
-  stockFromFeature,
   type TextFeatureData,
 } from '../types/project'
 import type {
@@ -64,7 +62,6 @@ import type {
   Tool,
 } from '../types/project'
 import type { OpenProfileEndpoint } from './types'
-import { convertProjectUnits } from '../utils/units'
 import { convertLength } from '../utils/units'
 import {
   featureHasClosedGeometry,
@@ -100,7 +97,7 @@ import {
   previewOffsetFeatures as previewOffsetFeaturesWithFactory,
   type DerivedFeatureGroup,
 } from './helpers/derivedFeatures'
-import { idNumericSuffix, nextPlacementSession, nextUniqueGeneratedId, syncIdCounter } from './helpers/ids'
+import { idNumericSuffix, nextUniqueGeneratedId, syncIdCounter } from './helpers/ids'
 import {
   angleToPoint,
   inferProfileOrientationAngle,
@@ -127,6 +124,7 @@ import { createOperationsSlice } from './slices/operationsSlice'
 import { createImportMergeSlice } from './slices/importMergeSlice'
 import { createProjectLifecycleSlice } from './slices/projectLifecycleSlice'
 import { createHistorySlice } from './slices/historySlice'
+import { createWorkpieceSlice } from './slices/workpieceSlice'
 import {
   propagateConstraintsOnTranslate,
   propagateConstraintsOnRotate,
@@ -2202,11 +2200,6 @@ export const useProjectStore = create<ProjectStore>((rawSet, get) => {
     future: [],
     transactionStart: null,
   },
-  setCreationTarget: (target) =>
-    set(() => ({
-      creationTarget: target,
-      pendingAdd: null,
-    })),
   ...createSelectionSlice(set, get, {
     cloneProject,
     normalizeProject,
@@ -2315,297 +2308,11 @@ export const useProjectStore = create<ProjectStore>((rawSet, get) => {
     projectsEqual,
     normalizeProject,
   }),
-
-  setOrigin: (origin) =>
-    set((s) => {
-      const nextProject = {
-        ...s.project,
-        origin,
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      }
-      if (projectsEqual(nextProject, s.project)) {
-        return {}
-      }
-      if (s.history.transactionStart) {
-        return { project: nextProject }
-      }
-      return {
-        project: nextProject,
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  startPlaceOrigin: () =>
-    set((s) => ({
-      pendingAdd: { shape: 'origin', session: nextPlacementSession() },
-      pendingMove: null,
-      pendingTransform: null,
-      sketchEditSession: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'origin' },
-        mode: 'feature',
-        hoveredFeatureId: null,
-        activeControl: null,
-      },
-    })),
-
-  placeOriginAt: (point) =>
-    set((s) => {
-      const nextProject = {
-        ...s.project,
-        origin: {
-          ...s.project.origin,
-          x: point.x,
-          y: point.y,
-        },
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      }
-      return {
-        project: nextProject,
-        pendingAdd: null,
-        pendingTransform: null,
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  // ── Stock ────────────────────────────────────────────────
-
-  setStock: (stock) =>
-    set((s) => {
-      const nextProject = {
-        ...s.project,
-        stock,
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      }
-      if (projectsEqual(nextProject, s.project)) {
-        return {}
-      }
-      if (s.history.transactionStart) {
-        return { project: nextProject }
-      }
-      return {
-        project: nextProject,
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  /**
-   * Set a feature as the stock source. The feature is removed from project.features
-   * and its geometry is used as the stock profile/thickness.
-   * Pass null to reset to rectangle stock (restores the feature to the tree).
-   *
-   * This is a single undo-able action that captures full before/after state.
-   */
-  setStockSourceFeature: (featureId: string | null) =>
-    set((s) => {
-      if (featureId === null) {
-        // Reset to rectangle stock
-        if (!s.project.stock.sourceFeatureId && !s.project.stock.sourceFeature) {
-          return {} // Already rectangle stock, no-op
-        }
-
-        const restoredFeature = s.project.stock.sourceFeature
-        if (!restoredFeature) return {}
-
-        const stockBounds = getStockBounds(s.project.stock)
-        const width = stockBounds.maxX - stockBounds.minX
-        const height = stockBounds.maxY - stockBounds.minY
-        const rectW = Math.max(width, 1)
-        const rectH = Math.max(height, 1)
-
-        const nextStock = {
-          ...s.project.stock,
-          profile: rectProfile(stockBounds.minX, stockBounds.minY, rectW, rectH),
-          sourceFeatureId: null as string | null | undefined,
-          sourceFeature: null as SketchFeature | null | undefined,
-        }
-
-        const nextProject = syncFeatureTreeProject({
-          ...s.project,
-          stock: nextStock,
-          features: [...s.project.features, restoredFeature],
-          meta: { ...s.project.meta, modified: new Date().toISOString() },
-        })
-
-        if (projectsEqual(nextProject, s.project)) {
-          return {}
-        }
-        if (s.history.transactionStart) {
-          return { project: nextProject }
-        }
-        return {
-          project: nextProject,
-          history: {
-            past: [...s.history.past, cloneProject(s.project)].slice(-100),
-            future: [],
-            transactionStart: null,
-          },
-        }
-      }
-
-      // Set a feature as stock source
-      const feature = s.project.features.find((f) => f.id === featureId)
-      if (!feature) return {}
-      if (!feature.sketch.profile.closed) return {} // Only closed profiles can be stock
-
-      // If another feature is already the stock source, restore it first
-      let features = s.project.features
-      let stock = { ...s.project.stock }
-
-      if (stock.sourceFeature && stock.sourceFeatureId) {
-        // Restore old source feature to features array
-        features = [...features, stock.sourceFeature]
-      }
-
-      // Remove the new source feature from features and tree
-      features = features.filter((f) => f.id !== featureId)
-      const featureTree = s.project.featureTree.filter(
-        (entry) => !(entry.type === 'feature' && entry.featureId === featureId)
-      )
-
-      // Build stock from feature
-      const newStock = stockFromFeature(feature)
-      stock = {
-        ...stock,
-        profile: newStock.profile,
-        thickness: newStock.thickness,
-        sourceFeatureId: feature.id,
-        sourceFeature: feature,
-      }
-
-      const nextProject = syncFeatureTreeProject({
-        ...s.project,
-        stock,
-        features,
-        featureTree,
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      })
-
-      if (projectsEqual(nextProject, s.project)) {
-        return {}
-      }
-      if (s.history.transactionStart) {
-        return { project: nextProject }
-      }
-      return {
-        project: nextProject,
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  /**
-   * Enter sketch edit mode for the stock source feature.
-   * Temporarily adds the source feature back to project.features and project.featureTree
-   * so that mutation actions (moveFeatureControl, insertFeaturePoint, etc.) can find and edit it.
-   * The feature is removed from features/tree on applySketchEdit (handled in selectionSlice).
-   */
-  enterStockSketchEdit: (featureId: string) =>
-    set((s) => {
-      const stock = s.project.stock
-      if (stock.sourceFeatureId !== featureId || !stock.sourceFeature) {
-        return {}
-      }
-
-      const feature = stock.sourceFeature
-
-      // Temporarily add the feature to features array and feature tree for editing
-      const nextProject = syncFeatureTreeProject({
-        ...s.project,
-        features: [...s.project.features, feature],
-        featureTree: [...s.project.featureTree, { type: 'feature' as const, featureId: feature.id }],
-      })
-
-      return {
-        project: nextProject,
-        pendingTransform: null,
-        pendingOffset: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: featureId,
-          selectedFeatureIds: [featureId],
-          selectedNode: { type: 'feature', featureId },
-          mode: 'sketch_edit',
-          sketchEditTool: null,
-          activeControl: null,
-        },
-        sketchEditSession: {
-          entityType: 'feature',
-          entityId: featureId,
-          snapshot: cloneProject(s.project),
-          pastLength: s.history.past.length,
-        },
-        pendingConstraint: null,
-      }
-    }),
-
-  setGrid: (grid) =>
-    set((s) => {
-      const nextProject = {
-        ...s.project,
-        grid,
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      }
-      if (projectsEqual(nextProject, s.project)) {
-        return {}
-      }
-      if (s.history.transactionStart) {
-        return { project: nextProject }
-      }
-      return {
-        project: nextProject,
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  setUnits: (units) =>
-    set((s) => {
-      if (s.project.meta.units === units) {
-        return {}
-      }
-
-      const convertedProject = convertProjectUnits(s.project, units)
-      const nextProject = {
-        ...convertedProject,
-        meta: { ...convertedProject.meta, modified: new Date().toISOString() },
-      }
-      if (projectsEqual(nextProject, s.project)) {
-        return {}
-      }
-      if (s.history.transactionStart) {
-        return { project: nextProject }
-      }
-      return {
-        project: nextProject,
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
+  ...createWorkpieceSlice(set, get, {
+    cloneProject,
+    projectsEqual,
+    syncFeatureTreeProject,
+  }),
 
   }
 })
